Updated documentation and representation of units.
