+++ essm/variables/units.py
@@ -25,6 +25,21 @@ SHORT_UNIT_SYMBOLS = {
     watt: var('W'),
 }
 
+
+def markdown(unit):
+    """Return markdown representaion of a unit."""
+    #FIXME consider (m/s)**(s/2)
+    facs = unit.factor_list()
+    str1 = ''
+    for term1 in facs:
+        op1 = term1[1]
+        if op1 == 1:
+            str1 = str(term1[0]) + ' ' + str1
+        else:
+            str1 += ' {0}$^{{{1}}}$ '.format(markdown(term1[0]), markdown(op1))
+    return str1
+
 __all__ = (
     'SHORT_UNIT_SYMBOLS',
+    'markdown',
 )
